Implemented the TestClass check. Reorganized the packages of the checks project
