+++ apex-checks/src/main/java/org/fundacionjala/enforce/sonarqube/apex/checks/unofficial/TestMethodCheck.java
@@ -21,66 +21,51 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.fundacionjala.enforce.sonarqube.apex.checks;
+package org.fundacionjala.enforce.sonarqube.apex.checks.unofficial;
 
-import java.util.regex.Pattern;
+import java.util.List;
 
 import com.sonar.sslr.api.AstNode;
-import com.sonar.sslr.api.Grammar;
 import org.sonar.api.server.rule.RulesDefinition;
 import org.sonar.check.Priority;
 import org.sonar.check.Rule;
-import org.sonar.check.RuleProperty;
 import org.sonar.squidbridge.annotations.ActivatedByDefault;
 import org.sonar.squidbridge.annotations.SqaleConstantRemediation;
 import org.sonar.squidbridge.annotations.SqaleSubCharacteristic;
-import org.sonar.squidbridge.checks.SquidCheck;
 
 import org.fundacionjala.enforce.sonarqube.apex.api.grammar.ApexGrammarRuleKey;
+import org.fundacionjala.enforce.sonarqube.apex.checks.Tags;
 
 /**
- * Verification of the name of the class.
+ * Verifies if a class contains test methods.
  */
 @Rule(
-        key = ClassNameCheck.CHECK_KEY,
+        key = TestMethodCheck.CHECK_KEY,
         priority = Priority.MAJOR,
-        name = "Class names should comply with a naming convention",
+        name = "Test methods incorrectly located",
+        description = "Test methods should be written in a test class",
         tags = Tags.CONVENTION
 )
 @SqaleSubCharacteristic(RulesDefinition.SubCharacteristics.READABILITY)
-@SqaleConstantRemediation("1min")
+@SqaleConstantRemediation("6min")
 @ActivatedByDefault
-public class ClassNameCheck extends SquidCheck<Grammar> {
+public class TestMethodCheck extends AnnotationMethodCheck {
+
+    /**
+     * Stores a message template.
+     */
+    public static final String MESSAGE = "The \"%s\" method corresponds to a test class.";
 
     /**
      * It is the code of the rule for the plugin.
      */
-    public static final String CHECK_KEY = "A1001";
-
-    /**
-     * The structure must have the name of the method.
-     */
-    private static final String DEFAULT = "^[A-Z_][a-zA-Z0-9]+$";
-
-    @RuleProperty(
-            key = "format",
-            defaultValue = "" + DEFAULT)
-    /**
-     * Stores the format for the rule.
-     */
-    public String format = DEFAULT;
-
-    /**
-     * Manages the pattern of rule.
-     */
-    private Pattern pattern = null;
+    public static final String CHECK_KEY = "A1007";
 
     /**
      * The variables are initialized and subscribe the base rule.
      */
     @Override
     public void init() {
-        pattern = Pattern.compile(format);
         subscribeTo(ApexGrammarRuleKey.CLASS_OR_INTERFACE_DECLARATION);
     }
 
@@ -92,10 +77,30 @@ public class ClassNameCheck extends SquidCheck<Grammar> {
      */
     @Override
     public void visitNode(AstNode astNode) {
-        String className = astNode.getFirstDescendant(ApexGrammarRuleKey.NAME).getTokenOriginalValue();
-        if (!pattern.matcher(className).matches()) {
-            getContext().createLineViolation(this,
-                    "Rename class \"{0}\" to match the regular expression {1}.", astNode, className, format);
+        AstNode modifierNode = null;
+        if (astNode.hasParent(ApexGrammarRuleKey.TYPE_DECLARATION)) {
+            modifierNode = astNode.getParent().getFirstChild(ApexGrammarRuleKey.MODIFIERS);
+        } 
+        if (!isAnnotation(modifierNode, IS_TEST)) {
+            List<AstNode> methods = astNode.getDescendants(ApexGrammarRuleKey.METHOD_DECLARATION);
+            methods.stream().forEach((method) -> {
+                AstNode parent = method.getParent();
+                AstNode firstChild = parent.getFirstDescendant(ApexGrammarRuleKey.MODIFIERS);
+                if (isTest(firstChild)) {
+                    getContext().createLineViolation(this, methodMessage(method), method);
+                }
+            });
         }
+    }
+
+    /**
+     * Returns the method message.
+     *
+     * @param astNode current node.
+     * @return the message.
+     */
+    private String methodMessage(AstNode astNode) {
+        AstNode method = astNode.getFirstDescendant(ApexGrammarRuleKey.METHOD_IDENTIFIER);
+        return String.format(MESSAGE, method.getTokenOriginalValue());
     }
 }
