+++ apex-checks/src/test/java/org/fundacionjala/enforce/sonarqube/apex/checks/TestMethodCheckTest.java
@@ -23,6 +23,7 @@
  */
 package org.fundacionjala.enforce.sonarqube.apex.checks;
 
+import org.fundacionjala.enforce.sonarqube.apex.checks.unofficial.TestMethodCheck;
 import java.io.File;
 
 import org.junit.Before;
