+++ apex-checks/src/test/resources/checks/testClassCheck.cls
@@ -1,0 +1,6 @@
+@anAnnotation
+@isTest
+@otherAnnotation
+public with sharing class SomeTestClass extends AClass {
+    
+}
