+++ apex-checks/src/test/resources/checks/testClassCheckBadName.cls
@@ -1,0 +1,4 @@
+@SomeOtherAnnotation
+public class SomeClassTest {
+	
+}
