+++ apex-checks/src/test/resources/checks/testClassCheckEnum.cls
@@ -1,0 +1,4 @@
+@isTest
+public enum SomeEnum {
+	SomeEnumValue, OtherEnumValue
+}
