+++ apex-checks/src/test/resources/checks/testClassCheckInterface.cls
@@ -1,0 +1,4 @@
+@isTest
+public interface SomeInterface {
+	
+}
